+++ packages/app/package.json
@@ -70,6 +70,7 @@
     "better-write-plugin-shortcuts": "^1.1.15",
     "better-write-plugin-theme": "^1.1.15",
     "better-write-plugin-voice-typing": "^1.1.15",
+    "better-write-plugin-webgl": "^1.1.15",
     "better-write-types": "^1.1.15",
     "compressorjs": "1.2.1",
     "destr": "1.2.2",
@@ -85,7 +86,7 @@
     "nprogress": "0.2.0",
     "ohmyfetch": "0.4.19",
     "pinia": "2.0.32",
-    "three": "0.151.2",
+    "three": "0.152.2",
     "vue": "3.2.47",
     "vue-directive-providers": "0.1.0",
     "vue-i18n": "9.2.2",
@@ -102,7 +103,7 @@
     "@rollup/plugin-inject": "4.0.4",
     "@types/file-saver": "2.0.5",
     "@types/jsdom": "16.2.14",
-    "@types/three": "0.150.0",
+    "@types/three": "0.151.0",
     "@typescript-eslint/eslint-plugin": "5.38.0",
     "@typescript-eslint/parser": "5.38.0",
     "@vitejs/plugin-vue": "4.1.0",
