+++ packages/plugin-webgl/.gitignore
@@ -1,0 +1,6 @@
+.vscode
+.DS_Store
+
+node_modules
+dist
+yarn-error.log
