+++ packages/plugin-webgl/.prettierrc.json
@@ -1,0 +1,6 @@
+{
+  "tabWidth": 2,
+  "singleQuote": true,
+  "semi": false,
+  "vueIndentScriptAndStyle": true
+}
