+++ packages/plugin-webgl/package.json
@@ -1,0 +1,45 @@
+{
+	"name": "better-write-plugin-webgl",
+	"version": "1.1.15",
+	"author": "Novout",
+	"license": "MIT",
+	"keywords": [
+		"vue",
+		"hook",
+		"text",
+		"editor"
+	],
+	"main": "dist/index.js",
+	"module": "dist/index.mjs",
+	"types": "dist/index.d.ts",
+	"scripts": {
+		"test": "echo \"Error: no test specified\"",
+		"build": "tsup",
+		"lint": "prettier --write src/"
+	},
+	"files": [
+		"dist/**/*",
+		"package.json",
+		"LICENSE",
+		"README.md"
+	],
+	"dependencies": {
+		"better-write-extension": "^1.1.15",
+		"better-write-plugin-core": "^1.1.15",
+		"better-write-types": "^1.1.15",
+		"vue-demi": "latest"
+	},
+	"devDependencies": {
+		"prettier": "2.5.1",
+		"rimraf": "3.0.2"
+	},
+	"peerDependencies": {
+		"@vue/composition-api": "^1.0.0-rc.1",
+		"vue": "^2.0.0 || >=3.0.0"
+	},
+	"peerDependenciesMeta": {
+		"@vue/composition-api": {
+			"optional": true
+		}
+	}
+}
