+++ packages/plugin-webgl/src/index.ts
@@ -1,0 +1,6 @@
+import { PluginTypes } from 'better-write-types'
+import { createPlugin } from 'better-write-plugin-core'
+import { BackgroundSet } from './background'
+
+export const WebGLPlugin = (): PluginTypes.Plugin =>
+  createPlugin({ name: 'webgl' }, [BackgroundSet])
