+++ packages/plugin-webgl/tsconfig.json
@@ -1,0 +1,38 @@
+{
+	"include": [
+		"src/**/*.ts"
+	],
+	"exclude": ["dist", "node_modules"],
+	"compilerOptions": {
+		"baseUrl": ".",
+		"rootDir": ".",
+		"outDir": "dist",
+		"sourceMap": false,
+		"noEmit": true,
+		"declaration": true,
+		"target": "esnext",
+		"module": "esnext",
+		"moduleResolution": "node",
+		"skipLibCheck": true,
+		"noUnusedLocals": true,
+		"strictNullChecks": true,
+		"noImplicitAny": true,
+		"noImplicitThis": true,
+		"noImplicitReturns": true,
+		"strict": true,
+		"isolatedModules": false,
+		"experimentalDecorators": true,
+		"resolveJsonModule": true,
+		"esModuleInterop": true,
+		"removeComments": false,
+		"strictPropertyInitialization": false,
+		"jsx": "preserve",
+		"lib": ["esnext", "dom"],
+		"types": ["node"],
+		"plugins": [
+			{
+				"name": "@vuedx/typescript-plugin-vue"
+			}
+		]
+	}
+}
