+++ packages/types/src/plugin/core.ts
@@ -77,6 +77,11 @@ export type PluginEmitterName =
   | 'plugin-oauth-login-with-provider'
   | 'plugin-oauth-logout'
   | 'plugin-oauth-delete'
+  | 'plugin-webgl-set-camera'
+  | 'plugin-webgl-loaded'
+  | 'call-landing-created'
+  | 'call-landing-mounted'
+  | 'call-landing-unmounted'
   | 'call-editor-created'
   | 'call-editor-mounted'
   | 'call-editor-unmounted'
